+++ CHANGELOG.md
@@ -12,7 +12,9 @@ changes.
 
 - **BREAKING** Switch to using `etcd` internally to establish a reliable L2 network
   - New run-time dependency onto `etcd` binary
-  - The same network options to `hydra-node` required (`--host`, `--port` and `--peer`) and need to match across the Hydra network.
+  - The peer network options to `hydra-node` (`--peer`) need to match across the Hydra network.
+  - `--host` and `--port` have been removed and we now have `--listen` and `--advertise`.
+  - `--listen` can be the same as the old `--host/--port` combined, but you may require `--advertise` to set your _public_ IP address and port combination, if you cannot bind to that on the local machine.
   - Adds `NetworkConnected` and `NetworkDisconnected` outputs which are most indicative of whether the L2 network is up or not.
   - Change `PeerConnected` and `PeerDisconnected` to indicate connectivity to `--peer` items and not the remote `node-id`.
   - Log outputs related to the network components changed significantly.